Lots of things - Networking working - FileSystem Working - Correct pinage - Resolved bugs with motors - Added things to webpage - More things i dont remember anymore
-- Lets go make some lidar readings now!!!!! --

+++ lib/Lidar/src/Lidar.cpp
@@ -9,7 +9,7 @@ void Lidar::setRPM(uint16_t RPM){
     _mode = MODE_RPM;
 }
 
-void Lidar::setMotor(uint8_t throttle){
+void Lidar::setMotor(uint16_t throttle){
     _target = throttle;
     _mode = MODE_THROTTLE;
 }
